pr reviews - lint errors
