+++ ArtemisKit/Sources/Messages/Views/SendMessageViews/SendMessageMentionContentView.swift
@@ -15,7 +15,7 @@ struct SendMessageMentionContentView: View {
     var body: some View {
         NavigationStack {
             let delegate = SendMessageMentionContentDelegate { [weak viewModel] mention in
-                if let selection = viewModel?.selection.wrappedValue {
+                if let selection = viewModel?.textSelection.wrappedValue {
                     switch selection.indices {
                     case .selection(let range):
                         viewModel?.text.insert(contentsOf: mention, at: range.upperBound)
